Fix Python3/unicode issues
This started off being an issue in that cell2code calls encode()
on a str value, turning it into a bytes value, and then tries appending
it to a str value. Python doesn't like this.

Fixing this required trying to work out what the correct behaviour
was. Which required adding tests.

In writing the tests I found some issues that also affected python 2.7.
In particular, under Python 2 we need a lot more u prefixes and handling
of quotes in strings was broken.

Some of these problems are fixed, however there are still places that
are broken.