Do not install_requires specific versions
We should not be requiring specific versions in setup.py. To do so
makes it difficult to upgrade a dependency in an application because it
then cause pip to complain that other dependencies require the old
version. It many cases the new version will work fine too - if not then
this is a bug that should get fixed rather then limiting the versions.

For more information see
https://packaging.python.org/discussions/install-requires-vs-requirements/