Support for col:col ranges
I ran into a similar issue as described in https://github.com/anthill/koala/issues/152
I took the maximum possible column range (16 384 columns) which is not the most efficient solution.
A more efficient way to implement this is to limit the number of columns, for instance by the used range. But this should be done in a coherent way over the sheets.