Fix resolve range (#186)
* New helper function to get the maximum dimensions of a sheet

* Resolve empty cells

When a range is defined containing cells that are empty, add those to the cellmap

* Add dependency

* Calculate the max dimensions

* Improve the speed and only run when there are addresses to add

* Add tests on sumifs and index match

Row (1:1) or col (A:A) style ranges
Ranges with jumps (e.g. empty cell in between)

* Add comments and style

* Improve comment

* Add docstring

* Indentation

* New helper function to get the maximum dimensions of a sheet

* Resolve empty cells

When a range is defined containing cells that are empty, add those to the cellmap

* Add dependency

* Calculate the max dimensions

* Improve the speed and only run when there are addresses to add

* Add tests on sumifs and index match

Row (1:1) or col (A:A) style ranges
Ranges with jumps (e.g. empty cell in between)

* Add comments and style

* Improve comment

* Add docstring

* Indentation

* Requirements

* Ignore tox files

* Update requirements.txt

* Update requirements.txt

* Attempt at fixing package issue on tox

* Update .travis.yml

* Update requirements.txt

* Improve tox file

* Fix test

* Fix test

With the ranges, this cell already exists so take another

* Don't install requirements from Travis to save time